refactor(instance-ai): Extract setup wizard into composables and utils (no-changelog)
Break down the 2198-line InstanceAiWorkflowSetup.vue into focused modules:

- instanceAiWorkflowSetup.utils.ts: types, constants, pure functions
- useCredentialTesting: credential test logic with store-derived state
- useCredentialGroupSelection: shared credential group state and mutations
- useSetupCards: card computation, display grouping, completion logic
- useSetupCardParameters: parameter value state and build-for-apply logic
- useSetupActions: apply, test-trigger, and later action flows

Parent component script reduced from ~1492 to ~455 lines.
All 15 existing tests pass unchanged.
