+++ packages/frontend/editor-ui/src/features/ai/instanceAi/instanceAiWorkflowSetup.utils.ts
@@ -1,0 +1,104 @@
+import type { InstanceAiWorkflowSetupNode } from '@n8n/api-types';
+import type { INodeProperties } from 'n8n-workflow';
+import { isResourceLocatorValue } from 'n8n-workflow';
+import type { INodeUi } from '@/Interface';
+
+// ---------------------------------------------------------------------------
+// Types
+// ---------------------------------------------------------------------------
+
+export interface SetupCard {
+	id: string;
+	credentialType?: string;
+	nodes: InstanceAiWorkflowSetupNode[];
+	isTrigger: boolean;
+	isFirstTrigger: boolean;
+	isTestable: boolean;
+	credentialTestResult?: { success: boolean; message?: string };
+	isAutoApplied: boolean;
+	hasParamIssues: boolean;
+}
+
+export interface SetupCardGroup {
+	parentNode: InstanceAiWorkflowSetupNode['node'];
+	parentCard?: SetupCard;
+	subnodeCards: SetupCard[];
+}
+
+export type DisplayCard =
+	| { type: 'single'; card: SetupCard }
+	| { type: 'group'; group: SetupCardGroup };
+
+// ---------------------------------------------------------------------------
+// Constants
+// ---------------------------------------------------------------------------
+
+export const HTTP_REQUEST_NODE_TYPE = 'n8n-nodes-base.httpRequest';
+export const HTTP_REQUEST_TOOL_NODE_TYPE = 'n8n-nodes-base.httpRequestTool';
+
+export const NESTED_PARAM_TYPES = new Set([
+	'collection',
+	'fixedCollection',
+	'resourceMapper',
+	'filter',
+	'assignmentCollection',
+]);
+
+// ---------------------------------------------------------------------------
+// Pure functions
+// ---------------------------------------------------------------------------
+
+export function credGroupKey(req: InstanceAiWorkflowSetupNode): string {
+	const credType = req.credentialType!;
+	const isHttpRequest =
+		req.node.type === HTTP_REQUEST_NODE_TYPE || req.node.type === HTTP_REQUEST_TOOL_NODE_TYPE;
+	if (isHttpRequest) {
+		const url = String(req.node.parameters.url ?? '');
+		if (url.startsWith('=')) {
+			return `${credType}:http:expr:${req.node.name}`;
+		}
+		return `${credType}:http:${url}`;
+	}
+	return credType;
+}
+
+/** Check if a parameter value is meaningfully set (not empty, null, or an empty resource locator). */
+export function isParamValueSet(val: unknown): boolean {
+	if (val === undefined || val === null || val === '') return false;
+	if (isResourceLocatorValue(val)) {
+		return val.value !== '' && val.value !== null && val.value !== undefined;
+	}
+	return true;
+}
+
+export function isNestedParam(p: INodeProperties): boolean {
+	return NESTED_PARAM_TYPES.has(p.type) || p.typeOptions?.multipleValues === true;
+}
+
+export function toNodeUi(setupNode: InstanceAiWorkflowSetupNode): INodeUi {
+	return {
+		id: setupNode.node.id,
+		name: setupNode.node.name,
+		type: setupNode.node.type,
+		typeVersion: setupNode.node.typeVersion,
+		position: setupNode.node.position,
+		parameters: setupNode.node.parameters as INodeUi['parameters'],
+		credentials: setupNode.node.credentials as INodeUi['credentials'],
+	} as INodeUi;
+}
+
+/** True when this card only has a trigger (no credentials and no param work) */
+export function isTriggerOnly(
+	card: SetupCard,
+	cardHasParamWork: (c: SetupCard) => boolean,
+): boolean {
+	return card.isTrigger && !card.credentialType && !cardHasParamWork(card);
+}
+
+/** Use credential icon when it's a credential card */
+export function shouldUseCredentialIcon(
+	card: SetupCard,
+	cardHasParamWork: (c: SetupCard) => boolean,
+): boolean {
+	return !!card.credentialType && !isTriggerOnly(card, cardHasParamWork);
+}
